Updates to be able to build this again!
Because of our infrastructure, I have not updated it properly to build in newer
Java version (mostly, this requires search/replace for 'javax' with 'jakarta').

Ideally this would move to a Maven-based build system but I don't have time.

I have also made a slight feature addition, which is that if you set the
latex-executable parameter to blank, the check servlet will not run that check.
There is currently still no way to disable the MathJax.node check.